add basic BE setup
