--- server/Model/index.js
@@ -1,3 +1,0 @@
-const User = require('./User')
-
-module.exports = User
